Fix #793: Allow to fully disable unpinning
As raised in #793, sometimes the user wants to simply disallow unpinning
altogether in a Cluster peer so that no content can be removed from IPFS
through cluster, even when crendentials are compromised (including access to
RPC endpoint).

This introduces a disable_unpinning parameter for the ipfshttp
connector. There is no way that a cluster will effecitvely unpin something if
the connector refuses to ask ipfs.